# Batch canvas draw operations and minimize redraws on map pans for the polygon and polyline layers to significantly improve performance.