… CHANGELOG.md
… example/lib/pages/polyline.dart
… pubspec.yaml
… test/layer/polygon_layer_test.dart
… test/layer/polyline_layer_test.dart