The batching is based on @MooNag 's proposal.

Based on a makeshift benchmark and my phone, I'm seeing:

* On the raster thread:
  * max: 2521ms => 8ms  (314x improvement)
  * avg: 1219ms => 6ms  (202x improvement)
* On the UI thread:
  * max: 49.6ms => 3.6ms (12.7x improvement)
  * avg: 19.2ms => 2.0ms (8.6x improvement)

* Before: 
![image](https://user-images.githubusercontent.com/111986/216788198-516f9fe1-9487-4ff3-97ad-b220c22af9d2.png)
* After: 
![image](https://user-images.githubusercontent.com/111986/216788215-34937987-a3c4-47a1-937e-ec9f95a43d7a.png)