@MooNag 2023-02-07
Reviewed this and tested the polygon layer in a a second app. I have made some improvements for performance and code quality. Ill try and commit them as a suggestion or a PR to your branch (not sure the exact right way to do that yet so i might cause some chaos. I might need to just do it as a review).

There is a chance that this will have an issue with culling enabled and caching the layer render, but i was not able to get it to consistently occur without toggling off isComplex in the CustomPaint.

Performance test in my real world app:
FM 3.1.0      - 15 FPS
This PR        - 42 FPS
My old PR   - 41 FPS

Id be happy to submit an approving review after taking a look at my recommendations.