@ignatz 2023-02-07
> Is there a specific reason we are using zoom and rotation in the shouldRepaint functions vs just using map.hashCode? We dont have to change it, it just seems like it would be cleaner code with the same results and less susceptible to breaking in the future.

I'm not sure I understand the suggestion of using map.hashCode correctly. shouldRepaint tells the repaint boundary on when to invalidate the cash, i.e. when two Painter objects should be considered visually different. If by map.hashCode you mean the FlutterMapState's dart-Object hashCode (which doesn't seem to be overridden), i'm not sure how that would achieve this.