@ignatz 2023-02-09
> Just going to have 1 more person check since its a fairly large rendering change and we should get it merged in soon. Please send me a ping if we don't get it merged in by next Friday (Feb 17th)!

That makes sense. I appreciate all the eyes we can get. Given the testing situation, I'm certainly not convinced that we're not introducing any new edge cases. I'd not be surprised if something bubbles up and we'll have to iterate. The less we need to, the better. Thanks