@ignatz 2023-02-16
> As far as I can see, everything is OK. As you said, there might be issues further down the road, but I think everything is pretty good.
> 
> In terms of the actual performance gain on my device, the UI thread is better, but the raster thread is the same, if not a little worse (@MooNag).
> 
> (master vs fast_poly in that order):

Thank you for taking the time to actually validate. Unfortunately for you, your observation is very much expected. The performance should be the same. The two optimizations kick in when you have:
 * polylines with the same paint property (bigger optimization)
 * and independently when you pan (smaller optimization)

In your case you're having few polylines with different properties + you're rotating/zooming. If you're still up for further validation, consider drawing some thousands polylines with the same color. Sorry, I should have been clearer..