@JaffaKetchup 2023-02-16
Ah no worries. As far as I'm concerened, two people saying performance has improved is enough, and this shows that normal performance has remained the ~same.
Apologies for taking so long to get this through! @MooNag's recently been invited to the repo with write perms, so he's going to merge when he's ready.