@MooNag 2023-02-16
Just checked the video, the performance regression seems insignificant enough. We will merge and watch for any issues!